requirements.txt
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
streamlit==1.33.0
python-dotenv==1.0.1
requests==2.31.0
scipy==1.13.0